+++ components/auth/index/AuthLinks.tsx
@@ -1,7 +1,7 @@
 import Link from 'next/link';
 import React from 'react';
 import styled from 'styled-components';
-import theme from '../../styles/theme';
+import theme from '../../../styles/theme';
 
 function AuthLinks() {
   return (
