+++ hooks/common/index.tsx
@@ -1,0 +1,5 @@
+import router from 'next/router';
+
+export const handleGoBack = () => {
+  router.back();
+};
