+++ pages/auth/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import AuthImage from '../../components/auth/AuthImage';
-import AuthLinks from '../../components/auth/AuthLinks';
+import AuthImage from '../../components/auth/index/AuthImage';
+import AuthLinks from '../../components/auth/index/AuthLinks';
 
 function Auth() {
   return (
