+++ types/signup/index.tsx
@@ -1,0 +1,3 @@
+export interface pageTitleProps {
+  pageTitleText: string;
+}
